Error reporting from asynchroneous operations
We now have a mechanism where errors are cloned, stored to a vector,
and only turned into a v8 Array back on the main thread.

We now use RAII to manage calls to xmlSetStructuredErrorFunc both for the
synchroneous and the asynchroneous case, setting the function up in the
constructor and resetting it in the destructor.

Instead of calling the push function for the arrays, we simply set elements
by index.  This simplifies code a lot, and should help performance as well.